Add files via upload
This is my capstone project for Machine Learning Engineer Nanodegree at Udacity. I build two models here for classifying dog breeds. It is an exploration and assessment of deep CNN for image classification problem. The pipeline I built can be used within a web or mobile app to process real-world, user-supplied images. Given an image of a dog, the algorithm will identify an estimate of the canine’s breed. If given an image of a human, the code will identify the resembling dog breed.